world_model_functionsをクラス毎にファイルわけ
