+++ aisaac/scripts/decision_maker.py
@@ -1,188 +1,10 @@
 # !/usr/bin/env  python
 # coding:utf-8
+
 import math
 import numpy as np
-import entity
-import sys
 import time
-import concurrent.futures
-from sklearn.cluster import KMeans
-import rospy
-#from consai_msgs.msg import VisionPacket, VisionIDList
-from consai_msgs.msg import robot_commands
-from consai_msgs.msg import Pose
-from nav_msgs.msg import Odometry
-from geometry_msgs.msg import Quaternion
-import tf
-from consai_msgs.msg import RefereeTeamInfo
-from std_msgs.msg import Int8
-import struct
-import serial
-import time
-#from aisaac.srv import Kick
-from aisaac.msg import Status
 
-WORLD_LOOP_RATE = 100.
-
-class WorldState:
-    def __init__(self, objects):
-        self.robot = objects.robot
-        self.robot_total = objects.robot_total
-
-        """---devisionの選択---"""
-        self.devision = "A"
-        #self.devision = "B"
-
-        """---フィールドとロボットのパラメータ---"""
-        self.robot_r = objects.robot[0].size_r
-        if self.devision == "A":
-            self.field_x_size = 12.
-            self.field_y_size = 9.
-        elif self.devision == "B":
-            self.field_x_size = 9.
-            self.field_y_size = 6.
-        self.field_x_min = - self.field_x_size / 2.
-        self.field_x_max = self.field_x_size / 2.
-        self.field_y_min = - self.field_y_size / 2.
-        self.field_y_max = self.field_y_size / 2.
-        self.goal_y_size = 1.2
-        self.goal_y_min = -self.goal_y_size / 2.
-        self.goal_y_max = self.goal_y_size / 2.
-
-        """---team cloorと攻撃方向---"""
-        #self.color = "Yellow"
-        self.color = "Blue"
-        self.goal_direction = "Right"
-        #self.goal_direction = "Left"
-
-        """positionの割り振り(動的にPositionが切り替わるのは多分大事、現状使っていない)"""
-        if self.robot_total == 8:
-            self.robot[0].position = "GK"
-            self.robot[1].position = "LCB"
-            self.robot[2].position = "RCB"
-            self.robot[3].position = "LSB"
-            self.robot[4].position = "RSB"
-            self.robot[5].position = "LMF"
-            self.robot[6].position = "RMF"
-            self.robot[7].position = "CF"
-        elif self.robot_total == 4:
-            self.robot[0].position = "GK"
-            self.robot[1].position = "LCB"
-            self.robot[2].position = "CCB"
-            self.robot[3].position = "RCB"
-
-
-class Referee:
-    def __init__(self, world_state):
-        self.world_state = world_state
-        """---refereeからの指示をもとにどのループを回すかの指標---"""
-        self.referee_branch = None
-
-        """refereeから受信する情報"""
-        self.command = None
-        self.stage = None
-        self.teaminfo = None
-
-    """---Refereeから現在のcommandをもらう---"""
-    def command_callback(self, msg):
-        self.command = str(msg)
-
-    """---Refereeから現在のstageをもらう---"""
-    def stage_callback(self, msg):
-        self.stage = str(msg)
-
-    """---Refereeから現在のteaminfoをもらう---"""
-    def teaminfo_callback(self, msg):
-        self.teaminfo = str(msg)
-
-    """---Refereeからの指示にしたがって行動指針を決定する---"""
-    def referee_branch_decision(self):
-        """
-        branch = 0 -> STOP
-        branch = 1 -> far from ball
-        branch = 2 -> normal
-        branch = 3 -> kickoff
-        """
-        #   print(self.command)
-        if self.command == "data: 0":
-            # Robots must keep 50 cm from the ball.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 1":
-            # A prepared kickoff or penalty may now be taken.
-            self.referee_branch = "STOP"
-        elif self.command == "data: 2":
-            # The ball is dropped and free for either team.
-            self.referee_branch = "NORMAL_START"
-
-        elif self.command == "data: 3":
-            # FORCE_START = 3
-		    #The yellow team may move into kickoff position.
-            self.referee_branch = "NORMAL_START"
-        elif self.command == "data: 4":
-            # PREPARE_KICKOFF_YELLOW = 4;
-		    #The blue team may move into kickoff position.
-            self.referee_branch = "KICKOFF"
-        elif self.command == "data: 5":
-            # PREPARE_KICKOFF_BLUE = 5;
-		    # The yellow team may move into penalty position.
-            self.referee_branch = "DEFENCE"
-        elif self.command == "data: 6":
-            # PREPARE_PENALTY_YELLOW = 6;
-		    # The blue team may move into penalty position.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 7":
-            # PREPARE_PENALTY_BLUE = 7;
-		    # The yellow team may take a direct free kick.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 8":
-            # DIRECT_FREE_YELLOW = 8;
-		    # The blue team may take a direct free kick.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 9":
-            #DIRECT_FREE_BLUE = 9;
-		    # The yellow team may take an indirect free kick.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 10":
-            #INDIRECT_FREE_YELLOW = 10;
-		    # The blue team may take an indirect free kick.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 11":
-            # INDIRECT_FREE_BLUE = 11;
-		    # The yellow team is currently in a timeout.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 12":
-            # TIMEOUT_YELLOW = 12;
-		    # The blue team is currently in a timeout.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 13":
-            # TIMEOUT_BLUE = 13;
-		    # The yellow team just scored a goal.
-		    # For information only.
-		    # For rules compliance, teams must treat as STOP.
-            self.referee_branch = "HALT"
-        elif self.command == "data: 14":
-            # GOAL_YELLOW = 14;
-		    # The blue team just scored a goal.
-            self.referee_branch = "STOP"
-        elif self.command == "data: 15":
-            # GOAL_BLUE = 15;
-            # Equivalent to STOP, but the yellow team must pick up the ball and
-		    # drop it in the Designated Position.
-            self.referee_branch = "STOP"
-        elif self.command == "data: 16":
-            #BALL_PLACEMENT_YELLOW = 16;
-	        #Equivalent to STOP, but the blue team must pick up the ball and drop
-		    # it in the Designated Position.
-            self.referee_branch = "STOP"
-        elif self.command == "data: 17":
-            #BALL_PLACEMENT_BLUE = 17;
-            self.referee_branch = "STOP"
-
-        if self.world_state.color == "Blue":
-            if self.command == "data: 4":
-                self.referee_branch = "DEFENCE"
-            elif self.command == "data: 5":
-                self.referee_branch = "KICKOFF"
 
 class DecisionMaker:
     def __init__(self, world_state, objects, referee, status):
@@ -475,65 +297,3 @@ class DecisionMaker:
             #print self.status[best_id].pid_goal_pos_x
             #print self.status[best_id].pid_goal_pos_y
             #print self.status[best_id].pid_goal_theta
-
-
-
-class Mesh:
-    def __init__(self, world_state, objects):
-        self.world_state = world_state
-        self.robot_total = objects.robot_total
-        self.enemy_total = objects.enemy_total
-        self.robot = objects.robot
-        self.enemy = objects.enemy
-
-    """フィールドをメッシュ化する(クラスタリング用に書かれた)"""
-    def create_mesh(self):
-        self.mesh_size = 50
-        self.mesh_x_num = int(self.world_state.field_x_size / (2*self.mesh_size)) + 1
-        self.mesh_y_num = int(self.world_state.field_y_size / self.mesh_size) + 1
-        self.mesh_x = [ i * self.mesh_size for i in range(self.mesh_x_num)]
-        self.mesh_y = [- self.world_state.field_y_size / 2 + i * self.mesh_size for i in range(self.mesh_y_num)]
-        #print(self.mesh_x)
-        #print(self.mesh_y)
-        self.mesh_xy = []
-        for x in self.mesh_x:
-            for y in self.mesh_y:
-                self.mesh_xy.append([x, y])
-        return self.mesh_xy
-
-    """各メッシュの密度計算"""
-    def enemy_density(self):
-        mesh_xy = np.array(self.create_mesh())
-        distance = np.zeros([self.mesh_x_num * self.mesh_y_num])
-        enemy_position = []
-        for i in range(self.robot_total-1):
-            x, y, _ = self.enemy[i+1].get_current_position()
-            #print(x,y)
-            distance_ = (mesh_xy - np.array([x,y]))**2
-            distance += np.sqrt(distance_[:,0] + distance_[:,1])
-        mean_distance = distance / 7
-        return mean_distance.reshape([self.mesh_x_num, self.mesh_y_num])
-
-    """密度からk-meansでクラス境界を獲得"""
-    def space_clustering(self):
-        self.threshold = 3000
-        self.n_cluster = 3
-        mean_distance = self.enemy_density().reshape([self.mesh_x_num * self.mesh_y_num])
-        mesh_xy = np.array(self.mesh_xy)
-        plots = mesh_xy[mean_distance > self.threshold]
-        print(plots.shape)
-        cls = KMeans(self.n_cluster)
-        pred = cls.fit_predict(plots)
-
-        #"""
-        for i in range(self.n_cluster):
-            labels = plots[pred == i]
-            plt.scatter(labels[:, 0], labels[:, 1])
-            #"""
-            centers = cls.cluster_centers_
-
-        plt.scatter(centers[:, 0], centers[:, 1], s=100,facecolors='none', edgecolors='black')
-        plt.xlim(-self.world_state.field_x_size/2, self.world_state.field_x_size/2)
-        plt.ylim(-self.world_state.field_y_size/2, self.world_state.field_y_size/2)
-        plt.show()
-        #"""
