--- aisaac/scripts/robot_functions.py
@@ -1,17 +1,0 @@
-#!/usr/bin/env  python
-# coding:utf-8
-import itertools
-import math
-import rospy
-import numpy as np
-from consai_msgs.msg import Pose
-from consai_msgs.msg import robot_commands
-#from aisaac.srv import Kick
-from aisaac.msg import Status
-from nav_msgs.msg import Odometry
-from geometry_msgs.msg import Quaternion
-import tf
-import entity
-import matplotlib.pyplot as plt
-from matplotlib import animation
-import functions
