+++ aisaac/scripts/world_model_utils.py
@@ -1,0 +1,4 @@
+# !/usr/bin/env  python
+# coding:utf-8
+
+WORLD_LOOP_RATE = 100.
